Mysql support

+++ domain/src/main/kotlin/com/astrainteractive/astrarating/domain/api/IRatingAPI.kt
@@ -4,13 +4,13 @@ import com.astrainteractive.astrarating.domain.entities.tables.dto.UserDTO
 import com.astrainteractive.astrarating.domain.entities.tables.dto.UserAndRating
 import com.astrainteractive.astrarating.domain.entities.tables.dto.UserRatingDTO
 
-const val NON_EXISTS_KEY = -1L
+const val NON_EXISTS_KEY = -1
 interface IRatingAPI {
     suspend fun selectUser(playerName: String): UserDTO?
     suspend fun updateUser(user: UserDTO)
-    suspend fun insertUser(user: UserDTO): Long?
+    suspend fun insertUser(user: UserDTO): Int?
 
-    suspend fun insertUserRating(it: UserRatingDTO): Long?
+    suspend fun insertUserRating(it: UserRatingDTO): Int?
     suspend fun deleteUserRating(it: UserRatingDTO)
 
     suspend fun fetchUserRatings(playerName: String): List<UserAndRating>?
